Backend/src/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
psycopg2-binary==2.9.9
pyway==0.3.26
PyYAML==6.0.1
strip-ansi==0.1.1
tabulate==0.9.0
typing_extensions==4.11.0
Werkzeug==3.0.2
pyjwt==2.8.0